Move some Jinja stuff around
Formatting functions that were originally in cms.contest.formatting but
that were later moved to methods of Translation make more sense in the
global environment. The same holds in the opposite direction for
get_score_class.

Also, since now those filters can be used by a wider audience, make them
more tolerant (i.e., use sensible defaults if the context variables they
need aren't present).
